initial testing suite, without null guards

+++ Debug/debug_blink.h
@@ -6,17 +6,16 @@
  * to provide visual feedback during debugging and testing.
  */
 
-
-
-
 #ifndef DEBUG_BLINK_H
 #define DEBUG_BLINK_H
 
+#include "../GPIO/gpio.h"
+#include "../SysTick/systick.h"
 #include <stdint.h>
 #include <stdbool.h>
 
 //manipulates GPIO 25 (onboard LED) to toggle every blink_rate_ms miliseconds in a loop
-void blink_loop(uint8_t blink_rate_ms);
+void blink_loop(uint32_t blink_rate_ms);
 
 //manipulates GPIO 25 (onboard LED) to blink a number of times, then returns
 void blink_n_times(uint8_t blink_count);
